get mm in a good state to debug..

+++ sys/include/commonarch/paging.h
@@ -7,10 +7,10 @@
 #define PAGE_SIZE_2M    (0x200000)
 #define PAGE_SIZE_1G    (0x40000000)
 
-#define LVL4_INDEX(va)  (((uintptr_t)(va) >> 39) & 0x1FF)
-#define LVL3_INDEX(va)  (((uintptr_t)(va) >> 30) & 0x1FF)
-#define LVL2_INDEX(va)  (((uintptr_t)(va) >> 21) & 0x1FF)
-#define LVL1_INDEX(va)  (((uintptr_t)(va) >> 12) & 0x1FF)
+#define LVL4_INDEX(va)  (((uint64_t)(va) >> 39) & 0x1FF)
+#define LVL3_INDEX(va)  (((uint64_t)(va) >> 30) & 0x1FF)
+#define LVL2_INDEX(va)  (((uint64_t)(va) >> 21) & 0x1FF)
+#define LVL1_INDEX(va)  (((uint64_t)(va) >> 12) & 0x1FF)
 
 #define SIGN_EXTEND(va) ()
